feat (#262): create hooks using gqless
