+++ packages/react/package.json
@@ -14,7 +14,7 @@
     "prepublish": "run-s build",
     "generate-docs": "typedoc",
     "generate-graphql": "graphql-codegen",
-    "test": "jest",
+    "test": "jest --passWithNoTests",
     "test:watch": "jest --watch",
     "ts": "tsc -p .",
     "ts:watch": "tsc -p . --watch"
