+++ packages/react/src/index.ts
@@ -1,18 +1,1 @@
-/* eslint-disable @typescript-eslint/triple-slash-reference */
-/// <reference path="./types/wpgraphql.d.ts" />
-
-export {
-  usePosts,
-  useGeneralSettings,
-  UriInfo,
-  useUriInfo,
-  usePost,
-} from './hooks';
-
-export {
-  HeadlessProvider,
-  useApollo,
-  addApolloState,
-  QueriesConfig,
-  HeadlessProviderPageProps,
-} from './provider';
+export { client } from './client';
